requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch==2.3.1
fastapi
uvicorn[standard]
pydantic
requests
pillow
google-genai
transformers
python-multipart